Initial Commit
Why these changes are being introduced:
This is the beginning of a block of code that will automate the
lifecyle of our Terraform GitHub repositories, our Terraform Cloud
workspaces, and that integration.

How this addresses that need:
* Creates consistent Terraform repositories in GitHub (based off a
pre-existing template)
* Create workspaces in Terraform Cloud (one for each required
environment/stage)
* Create initial variable values in each workspace to easily track the
app name, the AWS OU and the environment/stage.

Side effects of this change:
None.